==> 01_Prototype/00_Prepare_data.ipynb <==
We are going to prepare a simple dataset file with the labels of our oranges dataset

import os
from pathlib import Path
import pandas as pd

DATA_DIR = 'DATASETS/oranges'

dataset = []
for path in Path(DATA_DIR).rglob('*.jpg'):
    dataset.append({
        "image": path.name,
        "label": str(path.parent).split("/")[-1]
    })

df = pd.DataFrame(dataset)
df.head()

df.to_csv(os.path.join(DATA_DIR, 'labels.csv'), index=False)
