==> 01_Prototype/01_Process-data-for-EDA.ipynb <==
# Install wandb
!pip install wandb --quiet

import os
import numpy as np
from skimage import io
import pandas as pd
from tqdm.notebook import tqdm
import params

import wandb

DATA_DIR = 'DATASETS/oranges'

def _create_table(image_paths, labels):
    """
    Create a table with the dataset
    
    Args:
    - image_paths (list->str): List of image paths
    - labels (list->str): List of labels corresponding to image_paths
    
    Returns:
    A wandb table
    """
    assert len(image_paths)==len(labels), "Number of images and labels different!"
      
    table = wandb.Table(columns=["File_Name", "Images", "Dataset", "Label"])
    
    for i, image_path in tqdm(enumerate(image_paths), total=len(image_paths)):
        image_label = labels[i]
        image = io.imread(os.path.join(DATA_DIR, image_label, image_path))
        
        table.add_data(
            image_path,
            wandb.Image(image),
            "None", # we don't have a dataset split yet
            image_label
        )

    return table

### START A NEW WANDB RUN

https://docs.wandb.ai/ref/python/run

A run in W&B is a unit of computation we want to track

Normally most of the runs are experiments

But in this case we are going to use W&B to upload data

We need to pass a `project` parameter. A project in W&B is just a collection of runs

The `entity` represents a team (if we work collaboratory)

We will indicate in the `job_type` argument this run is for upload data. This will help us to **organize our runs** in the dashboard

run = wandb.init(
    project=params.WANDB_PROJECT,
    entity=params.ENTITY,
    job_type="upload"
)

### CREATE AN ARTIFACT

https://docs.wandb.ai/ref/python/artifact

We will use WANDB Artifacts to version our datasets

artifact = wandb.Artifact(
    params.RAW_DATA_AT,  # The name of the artifact
    type="raw_data"  # Will help us to organize our information
)

We can think of an Artifact as a version folder containing our data

We can add files or folders to our Artifact

artifact.add_dir(DATA_DIR, name="images")
artifact.add_file(os.path.join(DATA_DIR, "labels.csv"), name="labels.csv")

-------

df = pd.read_csv(os.path.join(DATA_DIR, "labels.csv"))
df.head()

table = _create_table(df.image, df.label)

-------

Add the table to our Artifact with a name

artifact.add(table, "eda_table")

Log the Artifact to WANDB

run.log_artifact(artifact)

Finally finish our run. That will upload our information to WANDB servers

run.finish()
