==> 01_Prototype/02_Split.ipynb <==
# Install wandb
!pip install wandb --quiet

!wandb --version

import os
import wandb
import pandas as pd
import numpy as np

from sklearn.model_selection import StratifiedGroupKFold

import params

Initialise a run to track de split

run = wandb.init(
    project=params.WANDB_PROJECT,
    entity=params.ENTITY,
    job_type="data_split"
)

In the previous Notebook we saved our data to an Artifact

We will use it now and track the lineage of our dataset in this way

raw_data_at = run.use_artifact(
    'marioparreno/mlops-wandb-course/oranges:latest',
    type='raw_data'
)
artifact_dir = raw_data_at.download()

To create the splits we will need the data filenames and labels

We have already that information at our dataset we retrieved from the artifact

orig_dataset = pd.read_csv(os.path.join(artifact_dir, "labels.csv"))
orig_dataset.head()

split_df = orig_dataset.copy()

Now we are going to fill the `Stage` column to define the splits for training, validation and test

validation_frac = 0.2
test_frac = 0.2
# implicit train size 1 - validation_size - test_size

# Set initially all to train split
split_df['stage'] = 'train'

# Shuffle the data
split_df = split_df.sample(frac=1, random_state=42).reset_index(drop=True)

# get the verfy first validate_size% as validation
val_size = int(len(split_df)*validation_frac)
split_df.loc[:val_size, ['stage']] = 'validation'

# get the next test_size% as test
test_size = int(len(split_df)*test_frac)
split_df.loc[val_size:(val_size+test_size-1), ['stage']] = 'test'

split_df.stage.value_counts()

Finally we save locally the data split

split_df.to_csv('data_split.csv', index=False)

And log the data split by using an Artifact. Create the Artifact

processed_data_at = wandb.Artifact(
    params.PROCESSED_DATA_AT,
    type="split_data"
)

Add the data relevant to the splitted dataset: 
- The raw data (we could process it, etc) but as we are not modifying => omit
- The split information + The dataset (labels) information

# The split information
processed_data_at.add_file('data_split.csv', name="data_split.csv")

We are going to save the split information by using the Table object from W&B

data_split_table = wandb.Table(
    dataframe=split_df
)

processed_data_at.add(data_split_table, "eda_table_data_split")

Now we can log our Artifact and finish the run

run.log_artifact(processed_data_at)
run.finish()
